가상자산 연결 중

+++ app/src/main/java/com/example/kobac_app/data/model/PortfolioResponse.kt
@@ -136,8 +136,8 @@ data class PortfolioResponse(
     val cryptoList: List<Any>,
     
     @SerializedName("totalNetWorthKrw")
-    val totalNetWorthKrw: Long,
+    val totalNetWorthKrw: Double,
     
     @SerializedName("totalNetWorthUsd")
-    val totalNetWorthUsd: Long
+    val totalNetWorthUsd: Double
 )
